+++ src/context/index.ts
@@ -1,3 +1,3 @@
 export { Source } from "./source";
-export { Location } from "./location";
+export { Location, sliceLocation } from "./location";
 export { Context, ContentModel } from "./context";